+++ Tests/LinuxMain.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+import FDBarGuageTests
+
+var tests = [XCTestCaseEntry]()
+tests += FDBarGuageTests.allTests()
+XCTMain(tests)